+++ auth_methods_userpass.go
@@ -1,9 +1,13 @@
 package main
 
 import (
-	log "github.com/Sirupsen/logrus"
 	"strings"
 )
+
+type cleanupAuthMethodUserpassTask struct {
+	path     string
+	userList UserList
+}
 
 type UserList map[string]interface{}
 
@@ -24,7 +28,12 @@ func configureUserpassAuth(auth authMethod) {
 	}
 
 	userpassAddUsers(auth.Path, userList)
-	cleanupUserpassUsers(auth.Path, userList)
+
+	cleanupAuthMethodUserpassTask := cleanupAuthMethodUserpassTask{
+		path:     auth.Path,
+		userList: userList,
+	}
+	taskPromptChan <- cleanupAuthMethodUserpassTask
 }
 
 func userpassAddUsers(path string, userList UserList) {
@@ -43,8 +52,8 @@ func userpassAddUsers(path string, userList UserList) {
 	}
 }
 
-func cleanupUserpassUsers(path string, userList UserList) {
-	existing_users, _ := Vault.List("/auth/" + path + "users")
+func (c cleanupAuthMethodUserpassTask) run(workerNum int) bool {
+	existing_users, _ := Vault.List("/auth/" + c.path + "users")
 
 	for _, v := range existing_users.Data {
 		switch userpassUsers := v.(type) {
@@ -53,26 +62,28 @@ func cleanupUserpassUsers(path string, userList UserList) {
 			for _, userpassUser := range userpassUsers {
 				switch username := userpassUser.(type) {
 				case string:
-					if _, ok := userList[username]; ok {
-						log.Debug("Userpass user [" + path + username + "] exists in configuration, no cleanup necessary")
+					if _, ok := c.userList[username]; ok {
+						log.Debug("Userpass user [" + c.path + username + "] exists in configuration, no cleanup necessary")
 					} else {
-						log.Info("Userpass user [" + path + username + "] does not exist in configuration, prompting to delete")
-						if askForConfirmation("Delete Userpass user ["+path+username+"] [y/n]?: ", 3) {
-							_, err := Vault.Delete("/auth/" + path + "users/" + username)
+						log.Info("Userpass user [" + c.path + username + "] does not exist in configuration, prompting to delete")
+						if askForConfirmation("Delete Userpass user [" + c.path + username + "]") {
+							_, err := Vault.Delete("/auth/" + c.path + "users/" + username)
 							if err != nil {
-								log.Fatal("Error deleting Userpass user ["+path+username+"] ", err)
+								log.Fatal("Error deleting Userpass user ["+c.path+username+"] ", err)
 							}
-							log.Info("Userpass user [" + path + username + "] deleted")
+							log.Info("Userpass user [" + c.path + username + "] deleted")
 						} else {
-							log.Info("Leaving Userpass user [" + path + username + "] even though it is not in config")
+							log.Info("Leaving Userpass user [" + c.path + username + "] even though it is not in config")
 						}
 					}
 				default:
-					log.Fatal("Issue parsing Userpass user from Vault [" + path + "]")
+					log.Fatal("Issue parsing Userpass user from Vault [" + c.path + "]")
 				}
 			}
 		default:
-			log.Fatal("Issue parsing Userpass users from Vault [" + path + "]")
+			log.Fatal("Issue parsing Userpass users from Vault [" + c.path + "]")
 		}
 	}
+
+	return true
 }