+++ secrets-engines.go
@@ -2,10 +2,24 @@ package main
 
 import (
 	"encoding/json"
-	log "github.com/Sirupsen/logrus"
-	VaultApi "github.com/hashicorp/vault/api"
 	"io/ioutil"
+
+	VaultApi "github.com/hashicorp/vault/api"
 )
+
+type syncSecretsEnginesTask struct {
+	secretsEnginesList map[string]SecretsEngine
+}
+
+type configureSecretsEngineTask struct {
+	secretsEngine SecretsEngine
+}
+
+type cleanupSecretsEnginesTask struct {
+	secretsEnginesList map[string]SecretsEngine
+}
+
+// type syncSecretsEngines struct {}
 
 type SecretsEngine struct {
 	Name         string
@@ -15,19 +29,33 @@ type SecretsEngine struct {
 	JustEnabled  bool // Flagged the first time a mount gets enabled
 }
 
-type SecretsEnginesList map[string]SecretsEngine
+func (s syncSecretsEnginesTask) run(workerNum int) bool {
 
-func SyncSecretsEngines() {
-
-	secretsEnginesList := SecretsEnginesList{}
+	// Decrement waitgroup counter when we're done
+	defer wg.Done()
 
 	log.Info("Syncing Secrets Engines")
-	GetSecretsEngines(secretsEnginesList)
-	ConfigureSecretsEngines(secretsEnginesList)
-	CleanupSecretsEngines(secretsEnginesList)
+	s.secretsEnginesList = make(map[string]SecretsEngine)
+	s.Load()
+
+	for _, secretsEngine := range s.secretsEnginesList {
+		configureSecretsEngineTask := configureSecretsEngineTask{
+			secretsEngine: secretsEngine,
+		}
+		wg.Add(1)
+		taskChan <- configureSecretsEngineTask
+	}
+
+	// Add cleanup task
+	cleanupSecretsEnginesTask := cleanupSecretsEnginesTask{
+		secretsEnginesList: s.secretsEnginesList,
+	}
+	taskPromptChan <- cleanupSecretsEnginesTask
+
+	return true
 }
 
-func GetSecretsEngines(secretsEnginesList SecretsEnginesList) {
+func (s syncSecretsEnginesTask) Load() {
 	files, err := ioutil.ReadDir(Spec.ConfigurationPath + "/secrets-engines/")
 	if err != nil {
 		log.Warn("No secrets engines found: ", err)
@@ -57,60 +85,63 @@ func GetSecretsEngines(secretsEnginesList SecretsEnginesList) {
 				}
 			}
 
-			secretsEnginesList[se.Path] = se
+			s.secretsEnginesList[se.Path] = se
 		}
 	}
 }
 
-func ConfigureSecretsEngines(secretsEnginesList SecretsEnginesList) {
-	for _, secretsEngine := range secretsEnginesList {
+func (c configureSecretsEngineTask) run(workerNum int) bool {
 
-		// Check if mount is enabled
-		existing_mounts, _ := VaultSys.ListMounts()
-		if _, ok := existing_mounts[secretsEngine.Path]; ok {
+	// Decrement waitgroup counter when we're done
+	defer wg.Done()
 
-			// We don't need to do any setup for identity backend
-			if secretsEngine.Path != "identity/" {
-				if existing_mounts[secretsEngine.Path].Type != secretsEngine.MountInput.Type {
-					log.Fatal("Secrets engine path ["+secretsEngine.Path+"] exists but doesn't match type; ", existing_mounts[secretsEngine.Path].Type, "!=", secretsEngine.MountInput.Type)
-				}
-				log.Debug("Secrets engine path [" + secretsEngine.Path + "] already enabled and type matches, tuning for any updates")
+	// Check if mount is enabled
+	existing_mounts, _ := VaultSys.ListMounts()
+	if _, ok := existing_mounts[c.secretsEngine.Path]; ok {
 
-				// Update the MountConfigInput description to match the MountInput description
-				// This is needed because of the way creating new mounts differs from existing ones?
-				secretsEngine.MountInput.Config.Description = &secretsEngine.MountInput.Description
-
-				err := VaultSys.TuneMount(secretsEngine.Path, secretsEngine.MountInput.Config)
-				if err != nil {
-					log.Fatal("Error tuning secrets engine path ["+secretsEngine.Path+"]", err)
-				}
+		// We don't need to do any setup for identity backend
+		if c.secretsEngine.Path != "identity/" {
+			if existing_mounts[c.secretsEngine.Path].Type != c.secretsEngine.MountInput.Type {
+				log.Fatal("Secrets engine path ["+c.secretsEngine.Path+"] exists but doesn't match type; ", existing_mounts[c.secretsEngine.Path].Type, "!=", c.secretsEngine.MountInput.Type)
 			}
-		} else {
-			log.Debug("Secrets engine path [" + secretsEngine.Path + "] is not enabled, enabling")
-			err := VaultSys.Mount(secretsEngine.Path, &secretsEngine.MountInput)
+			log.Debug("Secrets engine path [" + c.secretsEngine.Path + "] already enabled and type matches, tuning for any updates")
+
+			// Update the MountConfigInput description to match the MountInput description
+			// This is needed because of the way creating new mounts differs from existing ones?
+			c.secretsEngine.MountInput.Config.Description = &c.secretsEngine.MountInput.Description
+
+			err := VaultSys.TuneMount(c.secretsEngine.Path, c.secretsEngine.MountInput.Config)
 			if err != nil {
-				log.Fatal("Error mounting secret type ["+secretsEngine.MountInput.Type+"] mounted at ["+secretsEngine.Path+"]; ", err)
+				log.Fatal("Error tuning secrets engine path ["+c.secretsEngine.Path+"]", err)
 			}
-			log.Info("Secrets engine type [" + secretsEngine.MountInput.Type + "] enabled at [" + secretsEngine.Path + "]")
-			secretsEngine.JustEnabled = true
 		}
-
-		if secretsEngine.Path == "identity/" {
-			log.Info("Configuring Identity backend ", secretsEngine.Path)
-			configureIdentitySecretsEngine(secretsEngine)
-		} else if secretsEngine.MountInput.Type == "aws" {
-			log.Info("Configuring AWS backend ", secretsEngine.Path)
-			ConfigureAwsSecretsEngine(secretsEngine)
-		} else if secretsEngine.MountInput.Type == "database" {
-			log.Info("Configuring database backend ", secretsEngine.Path)
-			ConfigureDatabaseSecretsEngine(secretsEngine)
-		} else {
-			log.Warn("Secrets engine types other than [aws] and [database] not currently configurable, please open PR!")
+	} else {
+		log.Debug("Secrets engine path [" + c.secretsEngine.Path + "] is not enabled, enabling")
+		err := VaultSys.Mount(c.secretsEngine.Path, &c.secretsEngine.MountInput)
+		if err != nil {
+			log.Fatal("Error mounting secret type ["+c.secretsEngine.MountInput.Type+"] mounted at ["+c.secretsEngine.Path+"]; ", err)
 		}
+		log.Info("Secrets engine type [" + c.secretsEngine.MountInput.Type + "] enabled at [" + c.secretsEngine.Path + "]")
+		c.secretsEngine.JustEnabled = true
 	}
+
+	if c.secretsEngine.Path == "identity/" {
+		log.Info("Configuring Identity backend ", c.secretsEngine.Path)
+		configureIdentitySecretsEngine(c.secretsEngine)
+	} else if c.secretsEngine.MountInput.Type == "aws" {
+		log.Info("Configuring AWS backend ", c.secretsEngine.Path)
+		ConfigureAwsSecretsEngine(c.secretsEngine)
+	} else if c.secretsEngine.MountInput.Type == "database" {
+		log.Info("Configuring database backend ", c.secretsEngine.Path)
+		ConfigureDatabaseSecretsEngine(c.secretsEngine)
+	} else {
+		log.Warn("Secrets engine types other than [aws] and [database] not currently configurable, please open PR!")
+	}
+
+	return true
 }
 
-func CleanupSecretsEngines(secretsEnginesList SecretsEnginesList) {
+func (c cleanupSecretsEnginesTask) run(workerNum int) bool {
 	existing_mounts, _ := VaultSys.ListMounts()
 
 	for path, mountOutput := range existing_mounts {
@@ -118,11 +149,11 @@ func CleanupSecretsEngines(secretsEnginesList SecretsEnginesList) {
 		// Ignore default mounts
 		// generic = old kv store
 		if !(mountOutput.Type == "system" || mountOutput.Type == "cubbyhole" || mountOutput.Type == "identity" || mountOutput.Type == "kv" || mountOutput.Type == "generic") {
-			if _, ok := secretsEnginesList[path]; ok {
+			if _, ok := c.secretsEnginesList[path]; ok {
 				log.Debug("Secrets engine [" + path + "] exists in configuration, no cleanup necessary")
 			} else {
 				log.Debug("Secrets engine [" + path + "] does not exist in configuration, prompting to delete")
-				if askForConfirmation("Secrets engine ["+path+"] does not exist in configuration. Delete [y/n]?: ", 3) {
+				if askForConfirmation("Secrets engine [" + path + "] does not exist in configuration. Delete?") {
 					err := VaultSys.Unmount(path)
 					if err != nil {
 						log.Fatal("Error deleting mount ", path, err)
@@ -134,4 +165,6 @@ func CleanupSecretsEngines(secretsEnginesList SecretsEnginesList) {
 			}
 		}
 	}
+
+	return true
 }